==> notebooks/data_cleaning/03_happiness_report_analysis.ipynb <==
## Analyze the World Happiness Report Dataset.

Dataset found [HERE](https://www.kaggle.com/ajaypalsinghlo/world-happiness-report-2021).

This notebook aims to look at the number of countries featured in the World Happiness Report dataset, and choose which countries to focus on in future analyses. We also look at the years, and choose which ones provide the cleanest dataset with the countries for further development.

# Dependencies.
import pandas as pd

# Read in the World Happiness Index for country names.
df = pd.read_csv('resources/whr/world-happiness-report.csv')
df = df[['Country name', 'year', 'Life Ladder']]
df.head()

# Look at the data we hope to predict - year 2021.
df_21 = pd.read_csv('resources/whr/world-happiness-report-2021.csv')
df_21 = df_21[['Country name', 'Ladder score']]
df_21.head()

# Print full list of country appearances.
print(df['Country name'].value_counts().to_string())

# Check country counts in each dataset.
print(len(df['Country name'].value_counts()))
print(len(df_21['Country name']))

# Find the countries missing from 2021 - we will probably drop these.
countries = df['Country name'].unique().tolist()
countries_21 = df_21['Country name'].tolist()

missing = []
for country in countries:
    if country not in countries_21:
        missing.append(country)
        
missing

# Pivot table to years and life ladder scores.
df_pv = df.pivot(index='Country name', columns='year', values='Life Ladder')
df_pv

# Drop the countries missing in 2021 dataset.
df_pv = df_pv.drop(labels=missing, axis=0)
df_pv

# Check quantity of countries in each year.
df_pv.count()

Based on the analysis so far, we would likely be best served at using the years 2010-2019, by collecting the countries that have values in every year.

# Get DataFrame with just these columns.
short_df = df_pv[[2010, 2011, 2012, 2013, 2014, 2015, 2016, 2017, 2018, 2019]]
short_df.head()

# Drop all rows that contain any null values.
short_df = short_df.dropna()
short_df

# List the final country list.
main_countries = short_df.index.tolist()
main_countries

# Check which countries don't exist in the 2020 column.
countries2020 = df_pv[2020].dropna().index.tolist()

for country in main_countries:
    if country not in countries2020:
        print(country)

# Check which countries don't exist in the 2021 dataset.
countries2021 = df_21['Country name'].tolist()

for country in main_countries:
    if country not in countries2021:
        print(country)

Every country exists in the 2021 dataset, so we are covered to use that as our model test.

# Prepare 2021 data for join.
df_21.set_index('Country name', inplace=True)
df_21.rename(columns={'Ladder score': 2021}, inplace=True)
df_21

# Combine the DataFrames and save the table.
full_df = short_df.join(df_21)
full_df.to_csv('countries_happiness_2010_2019_2021.csv')
full_df.head()
