==> notebooks/data_cleaning/04_berkeley_temperature_data_pull.ipynb <==
Using the Berkley Earth website, we gather data to get the average temperature in each of our countries in the month of June.

# Dependencies.
import pandas as pd
import requests
# For downloading files.
import urllib.request

# Get country dataset.
df = pd.read_csv('countries_happiness_2010_2019_2021.csv')
df.head()

# Get the list of countries.
countries = df['Country name'].tolist()
countries[:5]

# Get the textual data from Berkeley Earth, and save it for each country.
for country in countries:
    try:
        country_url = country
        if ' ' in country:
            country_url = country.replace(' ', '-')
        urllib.request.urlretrieve(f'http://berkeleyearth.lbl.gov/auto/Regional/TAVG/Text/{country_url.lower()}-TAVG-Trend.txt', f'be_text/{country}_temp.txt')
        print(f'{country} completed.')
    except:
        print(f'COUNTRY FAILED - {country}')
