==> notebooks/data_cleaning/05_be_temp_analysis.ipynb <==
We will read in the Berkeley Earth text files, scraped from their website, and pull out the temperature data for each country. This data will be put into CSV files to allow for visualization, and also organized for machine learning.

# Dependencies.
import pandas as pd
import re

# Read in country dataset to get country list.
countries_df = pd.read_csv('countries_happiness_2010_2019_2021.csv')
countries_df.head()

# Get the country values in a list.
countries = countries_df['Country name'].tolist()
countries[:5]

# Gather all the initial temperatures to add to main DataFrame.
init_temps = []

for country in countries:
    # Access the text file for each country.
    filename = f'be_text/{country}_temp.txt'
    with open(filename) as f:
        content = f.readlines()
    
    # Get the initial temperature the data is based upon.
    init_temp = float(re.search(r'\(C\): (\S+) +', content[50]).group(1))
    init_temps.append(init_temp)
    
    # Store every row of temperature data, splitting by whitespace.
    temp_data = []
    for row in content[71:]:
        sep = [x for x in re.split('\s+', row) if x]
        temp_data.append(sep)
    
    # Create a DataFrame to hold all the data.
    temp_df = pd.DataFrame(temp_data, columns=['Year',
                                               'Month',
                                               'Monthly Anomaly',
                                               'Monthly Uncertainty',
                                               'Annual Anomaly',
                                               'Annual Uncertainty',
                                               'Five Year Anomaly',
                                               'Five Year Uncertainty',
                                               'Ten Year Anomaly',
                                               'Ten Year Uncertainty',
                                               'Twenty Year Anomaly',
                                               'Twenty Year Uncertainty'])
    
    # Store the data in a CSV with name including the initial temperature.
    temp_df.to_csv(f'be_csv/{country}_{init_temp}_temps.csv', index=False)

# Gather country names and initial temperatures.
data_tuples = list(zip(countries, init_temps))
df_init_temps = pd.DataFrame(data_tuples, columns=['Country name', 'Initial Temperature'])
df_init_temps.set_index('Country name', inplace=True)
df_init_temps.head()

# Add initial temperatures to main DataFrame.
countries_df.set_index('Country name', inplace=True)
full_df = countries_df.join(df_init_temps)
full_df.head()

# Save main DataFrame.
full_df.to_csv('countries_happiness_init_temps.csv')

For ease of looking up further data from APIs and websites we also changed 'North Macedonia' to 'Macedonia', 'Taiwan Province of China' to 'Taiwan', and dropped 'Palestinian Territories'.
