==> notebooks/data_cleaning/06_make_full_temp_ds.ipynb <==
Create Master Dataset of countries with temperature data.

# Dependencies.
import pandas as pd

# Import countries.
countries_df = pd.read_csv('countries_happiness_init_temps.csv')
countries_df.head()

# Rename columns to not mix up new temperature columns.
countries_df.rename(columns={
    '2010': 'LL_2010',
    '2011': 'LL_2011',
    '2012': 'LL_2012',
    '2013': 'LL_2013',
    '2014': 'LL_2014',
    '2015': 'LL_2015',
    '2016': 'LL_2016',
    '2017': 'LL_2017',
    '2018': 'LL_2018',
    '2019': 'LL_2019',
    '2021': 'LL_2021',
}, inplace=True)
countries_df.head()

# Get lists of countries and initial temps.
countries = countries_df['Country name'].tolist()
init_temps = countries_df['Initial Temperature'].tolist()
print(countries[:5])
print(init_temps[:5])

# Choose list of years to take June temps from.
years = [2010, 2011, 2012, 2013, 2014, 2015, 2016, 2017, 2018, 2019]

# Store every June temperature in the years for every country.
all_temps = []

# Loop through every country.
for i in range(len(countries)):
    # Read in the temperature data as a DF.
    temp_df = pd.read_csv(f'be_csv/{countries[i]}_{init_temps[i]}_temps.csv')
    
    # Find only June temps in chosen years, and add the initial value to them.
    june_temps = (temp_df.loc[(temp_df['Year'].isin(years)) & (temp_df['Month'] == 6)]['Monthly Anomaly'] + init_temps[i]).round(3).to_list()
    
    # Add June values to big list to hold all temperatures for all countries.
    all_temps.append(june_temps)

# Set column names for temperatures.
temp_cols = ['2010_TEMP',
             '2011_TEMP',
             '2012_TEMP',
             '2013_TEMP',
             '2014_TEMP',
             '2015_TEMP',
             '2016_TEMP',
             '2017_TEMP',
             '2018_TEMP',
             '2019_TEMP']

# Create DataFrame of all June temps for all countries.
june_temps_df = pd.DataFrame(all_temps, index=countries, columns=temp_cols)
june_temps_df.to_csv('temps_.csv')
june_temps_df.head()

# Set index of countries DataFrame to the country names.
countries_df.set_index('Country name', inplace=True)
countries_df.head()

# Merge the two datasets.
full_df = countries_df.join(june_temps_df)
full_df.head()

# Save the full country and temps DataFrame to a CSV.
full_df.to_csv('MAIN_countries_happiness_temps_full.csv')
