==> notebooks/data_cleaning/10_alternate_table_layout.ipynb <==
## Reorienting the Dataset.

Changing the orientation of the dataset in case we can get further analysis out of it.

# Dependencies.
import pandas as pd
import numpy as np

# Dataset.
df = pd.read_csv('MAIN_countries_happiness_temps.csv', index_col='Country name')
df.head()

# Drop the initial temperature column.
df.drop(columns='Initial Temperature', inplace=True)

# Add empty value column for 2021 temperature.
df['2021_TEMP'] = np.nan

# View again.
df.head()

# Get column list.
cols = df.columns.tolist()
cols

# Get duplicate country list.
countries = df.index.tolist()
countries_long = np.repeat(countries, 11)
countries_long[:12]

# Get similar list of years for each country.
years = [2010, 2011, 2012, 2013, 2014, 2015, 2016, 2017, 2018, 2019, 2021]
years_long = np.tile(years, 100)
years_long[:12]

# Create index object.
indices = [countries_long, years_long]

# Create column pairs.
col_pairs = []
for i in range(int(len(cols) / 2)):
    col_pairs.append([cols[i], cols[i+11]])
col_pairs

data = []
# Loop through each country and column pair and save each.
countries = df.index.tolist()
for country in countries:
    for col_pair in col_pairs:
        x = df.loc[df.index==country][col_pair].values
        data.append([x[0][0], x[0][1]])
        
data[:20]

# Create the new DataFrame with two indices and the reordered data.
reordered_df = pd.DataFrame(data, index=indices, columns=['Life Ladder', 'Temperature'])
reordered_df.index.names = ['Country', 'Year']
reordered_df.head(20)

# Save the DataFrame.
reordered_df.to_csv('MAIN2_countries_happ_temps_reorder.csv')
