==> notebooks/data_cleaning/11_add_clean_water.ipynb <==
## Add Clean Water Data to Reordered Dataset

Using the cleaned dataset from Suleiman's branch, we add the water data to the reordered dataset.

# Dependencies.
import pandas as pd
import numpy as np

# Get water dataset.
water_df = pd.read_csv('resources/Cleaned_Water_Data.csv', index_col='COUNTRY (DISPLAY)')
water_df.drop(columns='Unnamed: 0', inplace=True)
water_df.head()

# Rename countries to match MAIN data.
water_df.rename(index={
    'United Republic of Tanzania': 'Tanzania',
    'United States of America': 'United States',
    'Viet Nam': 'Vietnam',
    'Venezuela (Bolivarian Republic of)': 'Venezuela',
    'United Kingdom of Great Britain and Northern Ireland': 'United Kingdom',
    'North Macedonia': 'Macedonia',
    'Bolivia (Plurinational State of)': 'Bolivia',
    'Republic of Moldova': 'Moldova',
    'Serbia': 'Kosovo',
    'Russian Federation': 'Russia',
    'Republic of Korea': 'South Korea'
}, inplace=True)

# Get reordered dataset.
df = pd.read_csv('MAIN2_countries_happ_temps_reorder.csv', index_col=['Country', 'Year'])
df.head()

# Create list of years and countries to look up in water data.
years = [2010, 2011, 2012, 2013, 2014, 2015, 2016, 2017, 2018, 2019, 2021]
countries = pd.read_csv('MAIN_countries_happiness_temps.csv', index_col='Country name').index.tolist()

# Store the water values.
water_values = []

# Loop through each country and year and pull data.
for country in countries:
    for year in years:
        try:
            vals = water_df.loc[(water_df.index == country) & (water_df['YEAR (DISPLAY)'] == year)]['Numeric'].tolist()[:3]
            
            # Find the usable value in the list.
            try:
                val = int(vals[1])
                val = vals[1]
            except:
                try:
                    val = int(vals[2])
                    val = vals[2]
                except:
                    val = int(vals[0])
                    val = vals[0]
            # Save value.
            water_values.append(val)
        except:
            water_values.append(np.nan)
            if year != 2021:
                print(f'COUNTRY ISSUE: {country}, {year}')
                
water_values[:12]

# Add water values.
df['Clean Water'] = water_values

# View DF.
df.head()

# Save DataFrame.
df.to_csv('M_country_happ_temp_water.csv')
