==> notebooks/data_cleaning/13_greenhouse_gas_country_name_corr.ipynb <==
## Fill in Missing Countries Greenhouse Gas Data.

Some of the countries are in the dataset, but not named the same as the main dataset, this aims to fix that and fill in missing data.

# Dependencies.
import pandas as pd

# Greenhouse gas cleaned data.
gg_df = pd.read_csv('resources/greenhouse_gasses_clean.csv', index_col='Country Name')
gg_df

# Get list of countries in GG data.
gg_countries = gg_df.index.unique().tolist()

# Get main list of countries.
df = pd.read_csv('MAIN_country_happ_temp_water.csv')
countries = df['Country'].unique().tolist()

# List the countries "missing" in Greenhouse Gas dataset.
[c for c in countries if c not in gg_countries]

# Find corrected names in dataset, and rename.
gg_df.rename(index={
    'Egypt, Arab Rep.': 'Egypt',
    'Kyrgyz Republic': 'Kyrgyzstan',
    'North Macedonia': 'Macedonia',
    'Russian Federation': 'Russia',
    'Slovak Republic': 'Slovakia',
    'Korea, Rep.': 'South Korea',
    'Venezuela, RB': 'Venezuela',
    'Yemen, Rep.': 'Yemen'
}, inplace=True)

# Merge the datasets - code from Najeeb.
merged_df = pd.merge(df, gg_df,  how='left', left_on=['Country', 'Year'], right_on = [gg_df.index, 'Date'])
clean_df = merged_df.drop(columns=['Date'])
final_df = clean_df.rename(columns={"Value": "Greenhouse Gas Emissions"})
final_df

# Save dataset.
final_df.to_csv('MAIN_happ_temp_water_gg.csv')
