==> notebooks/data_cleaning/14_adding_population_data.ipynb <==
## Add Country Populations to Dataset

Add the populations of each country in each year to the dataset.

# Dependencies.
import pandas as pd

# Greenhouse gas cleaned data.
pop_df = pd.read_csv('resources/populations.csv')
pop_df.head()

# Gather only needed columns.
pop_df = pop_df[['Country Name', '2010', '2011', '2012', '2013', '2014', '2015', '2016', '2017', '2018', '2019']]
pop_df.rename(columns={
    '2010': 2010,
    '2011': 2011,
    '2012': 2012,
    '2013': 2013,
    '2014': 2014,
    '2015': 2015,
    '2016': 2016,
    '2017': 2017,
    '2018': 2018,
    '2019': 2019
}, inplace=True)
pop_df.head()

# Set the Countries as the DataFrame index.
pop_df.set_index('Country Name', inplace=True)
pop_df.head()

# Stack column to index.
stack_df = pd.DataFrame(pop_df.stack(), columns=['Population'])
stack_df.rename_axis(index=['Country', 'Year'], inplace=True)
stack_df.head()

# Read in full dataset.
df = pd.read_csv('ds13_happ_temp_water_gg.csv', index_col=['Country', 'Year'])
df.head()

# Compare country lists.
countries = df.index.levels[0]
countries_pop = stack_df.index.levels[0]

for country in countries:
    if country not in countries_pop:
        print(country)

# Find corrected names in dataset, and rename.
stack_df.rename(index={
    'Egypt, Arab Rep.': 'Egypt',
    'Kyrgyz Republic': 'Kyrgyzstan',
    'North Macedonia': 'Macedonia',
    'Russian Federation': 'Russia',
    'Slovak Republic': 'Slovakia',
    'Korea, Rep.': 'South Korea',
    'Venezuela, RB': 'Venezuela',
    'Yemen, Rep.': 'Yemen'
}, inplace=True)

# Compare country lists again.
countries = df.index.levels[0]
countries_pop = stack_df.index.levels[0]

for country in countries:
    if country not in countries_pop:
        print(country)

# Merge the datasets.
merged_df = pd.merge(df, stack_df,  how='left', left_on=['Country', 'Year'], right_on = ['Country', 'Year'])
merged_df.head()

# Save DataFrame.
merged_df.to_csv('ds14_country_happ_temp_cw_gg_pop.csv')
