==> notebooks/data_cleaning/15_adding_pm25_data.ipynb <==
## Add PM2.5 to Dataset

From Jeffrey's branch adding the cleaned PM2.5 dataset to main dataset.

# Dependencies.
import pandas as pd

# Greenhouse gas cleaned data.
pm_df = pd.read_csv('resources/pm2.5_per_cubic_metre.csv', index_col=['Country', 'Year'])
pm_df.drop(columns=['Unnamed: 0', 'Code'], inplace=True)
pm_df.rename(columns={'pm_2.5_per_cubic_metre': 'PM2.5'}, inplace=True)
pm_df.head()

# Read in main dataset.
df = pd.read_csv('ds14_country_happ_temp_cw_gg_pop.csv', index_col=['Country', 'Year'])
df.head()

# Compare country lists.
countries = df.index.levels[0]
countries_pm = pm_df.index.levels[0]

for country in countries:
    if country not in countries_pm:
        print(country)

# Find corrected names in dataset, and rename.
pm_df.rename(index={
    'Chinese Taipei': 'Taiwan',
    'Serbia': 'Kosovo',
    'North Macedonia': 'Macedonia',
    'Slovak Republic': 'Slovakia',
    'Korea': 'South Korea',
    'Viet Nam': 'Vietnam'
}, inplace=True)
pm_df.head()

# Compare country lists again.
countries = df.index.levels[0]
countries_pm = pm_df.index.levels[0]

for country in countries:
    if country not in countries_pm:
        print(country)

# Merge the datasets.
merged_df = pd.merge(df, pm_df,  how='left', left_on=['Country', 'Year'], right_on = ['Country', 'Year'])
merged_df.head()

# Save dataset.
merged_df.to_csv('ds15_happ_temp_cw_gg_pop_pm.csv')
