==> notebooks/data_cleaning/19_add_co2_emissions.ipynb <==
## Add CO<sub>2</sub> Emission Data to Reordered Dataset

Using the cleaned dataset from Jeffrey's branch, we add the emission data to the reordered dataset.

# Dependencies.
import pandas as pd

# Get cleaned CO2 dataset.
co2_df = pd.read_csv('resources/annual_co2_2010-2019.csv', index_col=['Country', 'Year'])
co2_df.rename(columns={'Annual CO2 emissions': 'CO2 Emissions'}, inplace=True)
co2_df.head()

# Read in main dataset.
df = pd.read_csv('ds15_happ_temp_cw_gg_pop_pm.csv', index_col=['Country', 'Year'])
df.head()

# Compare country lists.
countries = df.index.levels[0]
countries_co2 = co2_df.index.levels[0]

for country in countries:
    if country not in countries_co2:
        print(country)

All countries featured in dataset, already cleaned by Jeffrey, simple to merge and create new CSV.

# Merge DataFrames.
merged_df = pd.merge(df, co2_df,  how='left', left_on=['Country', 'Year'], right_on = ['Country', 'Year'])
merged_df.head(20)

# Save dataset.
merged_df.to_csv('ds19_happ_temp_cw_gg_pop_pm_co2.csv')
