==> notebooks/data_cleaning/31_dataset_per_capita.ipynb <==
## Create a New Dataset with per Capita Values.

Defined as per-person, so we will take values and divide them by the population at the time for each country. This does not need to be performed on Life Ladder, Temperature, Clean Water, PM2.5, or Population. Also drop the changed columns to make a new dataset.

# Dependencies.
import pandas as pd

# Data.
df = pd.read_csv('ds19_happ_temp_cw_gg_pop_pm_co2.csv', index_col=['Country', 'Year'])
df.head()

# Make new columns, drop old ones.
df['pc Greenhouse Gas Emissions'] = df['Greenhouse Gas Emissions'] / df['Population']
df['pc CO2 Emissions'] = df['CO2 Emissions'] / df['Population']
df = df.drop(columns=['CO2 Emissions', 'Greenhouse Gas Emissions', 'Population'])
df.head()

# Save the dataset.
df.to_csv('ds31_all_per_capita_emissions.csv')
